refactor: rename `run` module to `execute` and move plan to top-level module
Also:
- consolidates `result` module into `execute` module.
- exposes plan and execute apis to top-level
