+++ benchmarks/haystack/src/main.rs
@@ -1,10 +1,6 @@
 use clap::Parser;
 use petname::Generator; // Trait needs to be in scope for `iter`.
-use spnl::{
-    Query,
-    run::{RunParameters, result::SpnlError, run},
-    spnl,
-};
+use spnl::{ExecuteOptions, Query, SpnlError, execute, spnl};
 
 type GeneratedNames = Vec<String>;
 #[derive(serde::Deserialize)]
@@ -136,7 +132,7 @@ async fn main() -> Result<(), SpnlError> {
     let system_prompt = r#"Your are an AI that responds to questions with a plain JSON array of strings such as ["a","b","c"] or ["x","y","z","w"] or ["hello","world"], no markdown or html or any other extra text"#;
     let user_prompt = "Tell me the names of the cats mentioned";
 
-    let program: Query = if chunk > 0 {
+    let query: Query = if chunk > 0 {
         let chunks: Vec<Query> = docs
             .chunks(chunk)
             .map(|chunk| chunk.to_vec())
@@ -175,14 +171,14 @@ async fn main() -> Result<(), SpnlError> {
 
     if !quiet {
         let mut stderr = ::std::io::stderr();
-        ptree::write_tree(&program, &mut stderr)?;
+        ptree::write_tree(&query, &mut stderr)?;
     }
 
     if dry_run {
         return Ok(());
     }
 
-    match run(&program, &RunParameters { prepare: None }).await? {
+    match execute(&query, &ExecuteOptions { prepare: None }).await? {
         Query::User(ss) => {
             // oof, be gracious here. sometimes the model wraps the
             // requested json array with markdown even though we asked
