+++ spnl/src/plan.rs
