--- spnl/src/run/result.rs
@@ -1,2 +1,0 @@
-pub type SpnlError = anyhow::Error;
-pub type SpnlResult = anyhow::Result<crate::Query>;
